requirements.txt
streamlit>=1.35.0
chromadb>=1.5.8
google-genai==1.66.0
pymupdf>=1.24.0
python-docx>=1.1.2
openpyxl>=3.1.2
python-dotenv>=1.0.1
langchain-text-splitters>=0.3.0